+++ src/main/kotlin/dsl/Skill.kt
@@ -1,0 +1,3 @@
+package dsl
+
+data class Skill(val name: String, val type: SkillType)
