+++ src/main/kotlin/dsl/SkillType.kt
@@ -1,0 +1,6 @@
+package dsl
+
+enum class SkillType {
+    SOFT,
+    HARD
+}
